!refactor(API): various changes to routes design - prefix all routes with v1 - adjust the domain entity route design to be plural, path args, and query string only for filters. trying to follow REST
